Update snmp functionality to include editing the postgresql config through snmpset
